refactor: package 수정
- 가독성을 위한 세부 패키지 제거
