+++ src/main/java/calculator/Application.java
@@ -3,10 +3,10 @@ package calculator;
 import calculator.adder.NumberAdder;
 import calculator.adder.NumberAdderImpl;
 import calculator.presentation.CalculatorController;
-import calculator.presentation.input.CalculatorInput;
-import calculator.presentation.output.CalculatorOutput;
-import calculator.presentation.parser.InputParser;
-import calculator.presentation.parser.RegexInputParser;
+import calculator.presentation.CalculatorInput;
+import calculator.presentation.CalculatorOutput;
+import calculator.presentation.InputParser;
+import calculator.presentation.RegexInputParser;
 import calculator.tokenizer.Delimiters;
 import calculator.tokenizer.RegexTokenizer;
 import calculator.tokenizer.Tokenizer;
