+++ src/main/java/calculator/presentation/InputParser.java
@@ -1,0 +1,6 @@
+package calculator.presentation;
+
+public interface InputParser {
+
+    CalculatorRequest parseToRequest(String input);
+}
