--- src/main/java/calculator/presentation/parser/InputParser.java
@@ -1,8 +1,0 @@
-package calculator.presentation.parser;
-
-import calculator.presentation.request.CalculatorRequest;
-
-public interface InputParser {
-
-    CalculatorRequest parseToRequest(String input);
-}
