+++ src/test/java/calculator/presentation/RegexInputParserTest.java
@@ -1,7 +1,5 @@
 package calculator.presentation;
 
-import calculator.presentation.parser.RegexInputParser;
-import calculator.presentation.request.CalculatorRequest;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
